pyproject.toml
[tool.poetry]
name = "pyasic-web"
version = "0.1.0"
description = ""
authors = ["UpstreamData <brett@upstreamdata.ca>"]
readme = "README.md"
packages = [{include = "pyasic_web"}]

[tool.poetry.dependencies]
python = "^3.8"
pyasic = "^0.33.9"
aiofiles = "^23.1.0"
fastapi = "^0.95.1"
websockets = "^11.0.3"
uvicorn = "^0.21.1"
itsdangerous = "^2.1.2"
Jinja2 = "^3.1.2"
sse-starlette = "^1.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
djlint = "^1.19.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
